requirements/base.txt
Django==1.8.14
Unipath==1.1
djangorestframework==3.4.7
sendgrid-django==4.0.0
Pillow==3.4.2
dj-database-url==0.4.1
gunicorn==19.6.0
whitenoise==3.2.2